refactor(lobe-chat): extract user bubble body into ThreadUserBody.tsx
ConversationThread.tsx held the whole transcript stack; the user-bubble
cluster (UserBodyText + UserPlainOrSkills + UserMessageBody, ~263 lines)
is self-contained: text/skill chips/quote cards/fold, plus the scheduled
and Remote IM pill headers.

Moved verbatim into lobe-chat/ThreadUserBody.tsx; only UserMessageBody
is exported (the row's sole consumer), the inner two stay module-local.
No behavior change — same memo components, same props.

ConversationThread.tsx: 3322 -> 3042 lines.
